Upgrade test to use artisan testing helper
